feat(ci): shuffle remote builders
Currently, when selecting a remote builder to run tests on, or build against, we
always select the one at top of the list. This isn't really taking full
advantage of our infrastructure, since we offer multiple machines for all
architectures currently.

This introduces a `shuf` before we select the builder.

Refs: flox/deltaops#375