docs(nvm.md): Improve utility of Flox command snippets (#115)
I solicited feedback from a former colleague, who made the very helpful
suggestion that each command that the user is meant to run should be
rendered in its own snippet that can be copied by clicking on the copy
to clipboard button. This makes the guide more user friendly and
prevents any issues that might arise when the reader misses a command
that's embedded in a block of text.