flox: Upgrade markdownlint and lychee
I want `gitignore` support from `markdownlint-cli2` but it wasn't
available in the version that was pinned by `python.version` in the
`toplevel` package group.

I've moved `lychee` to a new group at the same time because it felt
weird having a group for a single package and I can safely verify that
these don't introduce any breaking changes.