lint: Ignore gitignored dirs (command-reference)
We don't want to lint `docs/reference/command-reference` because the
files are already committed and copied from `flox/flox`. We may choose
to apply linting there separately.

We could use `!docs/reference/command-reference` in the glob patterns
but it feels more reliable to depend on the gitignore rules that we
already have in case the path changes at any time.