activation: Link to modes in manifest.toml
The content here has been moved and slightly modified to the new
`[options.activate]` section in `manifest.toml` and the `flox activate`
docs updated to reference that so that we don't have to maintain 4+
copies of the same content.

I've verified locally that the anchor link works but the new content
won't be present until flox/flox#2781 is released so this should be
merged at the same time.