feat(compose): Add remote to tutorial
Now that we support remote includes. I considered rewriting the
"creating" section to reference remote first but it considerably changes
the flow of the document and will need re-testing from the start, so for
now I've opted to add it into a separate section and explain the benefit
of using remotes.