feat(compose): Support remote includes (#212)
Now that we support remote includes. I considered rewriting the
"creating" section to reference remote first but it considerably changes
the flow of the document and will need re-testing from the start, so for
now I've opted to add it into a separate section and explain the benefit
of using remotes.

This should only be merged with the corresponding release of
https://github.com/flox/flox/pull/2952 ~~which isn't currently in
1.4.0-rcX~~

---------

Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>