concepts: Merge manifest concept into environments
The `/concepts/manifest` page had content that overlapped with both:

- `manifest.toml` which contains a more complete listing of all the
  currently supported sections.
- `environments` which describes the files stored in `.flox/env/`

So merge it with environments, introduce the basic format and how to
edit, then link out to `manifest.toml` for more details.

The old URL will be redirected using the functionality that was added in
91141a7. It's worth noting that this doesn't return a 301 response
because the redirect is performed client-side from static content:

    % http :8000/docs/concepts/manifest/
    HTTP/1.0 200 OK
    Content-Length: 1280
    Content-Type: text/html
    Date: Thu, 15 May 2025 09:35:46 GMT
    Server: WSGIServer/0.2 CPython/3.11.9

    <!doctype html>
    <html lang="en">
    <head>
        <meta charset="utf-8">
        <title>Redirecting...</title>
        <link rel="canonical" href="../environments/#files-manifest-toml">
        <script>var anchor=window.location.hash.substr(1);location.href="../environments/#files-manifest-toml"+(anchor?"#"+anchor:"")</script>
        <meta http-equiv="refresh" content="0; url=../environments/#files-manifest-toml">
    </head>
    <body>
    You're being redirected to a <a href="../environments/#files-manifest-toml">new destination</a>.
    <script>
    …

However it appears that search engines do treat this is a permanent
redirect (https://stackoverflow.com/a/19717455) and we can't redirect to
a new anchor using Netlify server-side redirects because they are a
client-side only concept (not sent as part of the request).

I don't think that we necessarily need to change the link from
the `flox init` templated manifest because the original URL is more
concise, even if it doesn't fit within the structure of the docs.