feat: rewrite customizing environments (#243)
This previously focused exclusively on shell hooks and used them to run
a service. This now shows a user how to leverage the hook, profile, and
vars sections to customize their shell environment in an activated
environment.

There is now a `lint` alias in the environment to make it easier to
remember the linter command.