chore: add customer section
This moves the existing "publishing" cookbook pages and creates a new
"known issues" page to a new "customer" section. This section isn't
linked in the sidebar, but can still be linked to if the URL is known.
Note that the pages in this section are still world-readable if you
know the URL.