docs(cookbooks): Add NEF examples
Add examples for all of the languages that we currently have in:

- https://github.com/flox/flox-build-examples

This required nesting the existing manifest build content under a new
higher-level heading to make way for Nix expression build content in
parallel.

I don't love how long this has made some of the pages, how hard it is
to find a complete example to get going, and how likely we are to keep
the examples up-to-date when they're not tested. I'd prefer to link out
to the other repo rather than duplicate the content here but "the team"
have a preference to keep the long form content here.

I also considered tabbing the content so that you can compare impure
manifest builds, pure manifest builds, and Nix expression builds in one
place, but it doesn't fit well with the narrative grouped content that
we have currently.