Generate llms.txt and doc-content.txt for AI tools
This commit adds the generations of llms.txt (more robust than what was there)
as well a single docs-content.txt that can be fed to an agent or answer engine
to set context of how to work with Flox.

This does add a Makefile (sorry) because I needed something to run multiple
steps when building the documentaiton site.

The README has been updated to explain Makefile usage (and it's completely optional).

The tools directory is also new, with a small python file to generate the AI indexing files.

The robots.txt that helps the crawlers know where to work is maintained in the
flox/floxwebsite repo.